elementary_ext: introduce mimetype cache
With efreet calls each time the icon resized we lost a lot of time and
the window widget caused a freeze for a short time. With the new cache
we just have one single call to a hash table, or in the worst case a
call to efreet (but only once for each mimetype).