elm_file_display: Listen for key events to pass to the search.
Notice that the handling of the view is horrible no arrow keys are
working, directorys are just selected, the old ones not unselected. Will
be the next step to work on that.

Also: right now its not shown on the ui what is searched for, will need
to add this to the theme.